chore(pkgs): executes a massive lockfile bump
this commit performs migrations of every package to its latest compatible or allowed version in the
workspace. it attempts to respect peer dependency versions as far as maintaining compatibility
between packages and not introducing any breaking changes. it will likely need considerable review,
and it would be advisable to clear all caches and reset nx daemons locally (yarn nx reset) and clear
cache in the nx cloud also, as well as removing node_modules locally, and `yarn nx repair` should be
run afterwards.