build: update node docker image to 18.16.0 and simplify workflow in .circleci/config.yml
This commit updates the node docker image used to build the project to version 18.16.0. It also simplifies the CircleCI workflow by removing unnecessary steps related to caching and restoring caches, and adding new parameters for nx/set-shas, allowing it to run with an on-hold workflow and specifying a main branch name. Finally, the production deployment is now filtered to run only on the main branch.