build: dist/apps/my-app-server feat: update yarn commands to npx in .circleci/config.yml
This commit updates the yarn commands in the deploy and build-app jobs in the .circleci/config.yml file to use npx instead of yarn. Specifically, it replaces "yarn --silent run" with "npx" for the nx commands in both jobs. This change simplifies the command syntax and ensures that the latest version of nx is used during the linting and building processes.