build: add yarn cache and restore cache steps for CircleCI
This commit adds yarn caching to the CircleCI workflow to speed up dependencies installation, and also adds restore_cache steps to restore the cache when running jobs. Additionally, two new jobs are added to the workflow to lint and build the app using Nx affected. Finally, artifact storing is added to store the resulting build output.