build: update yarn cache keys
This commit updates the keys used for caching dependencies. Specifically, the key prefix has been changed from "v1-dependency-cache" to "v3-dependency-cache". The new key is based on the checksum of the "yarn.lock" file rather than "package.json". This change ensures that the correct cache is used for each project and improves cache-hit rates. Additionally, this commit updates the cache key used for saving the cache.