feat: update cache key from v5 to v6 in CircleCI configuration and remove nx lint fix from package.json
This commit updates the cache key from 'v5' to 'v6' in the CircleCI configuration file, specifically for the 'yarn_cache' key. Additionally, it removes the 'lint:fix' script from the 'package.json' file, which was previously used to run linting and formatting fixes across the entire workspace. Instead, the 'yarn nx affected' command is now used to target only affected projects for linting.