Feature/plp api (#183)
* feat(server): created prototype plp api endpoint

* feat: update jewelry products

* feat: pdp for ER + Jewelry in prog

* feat: action feedback

* feat: remove commented out

* feat: scaffold basic pdp

* refactor: updated the main diamonds api from post to get

* linting on ci

* feat(nx): make nx aware of package.json scripts

* build: update node docker image to 18.16.0 and simplify workflow in .circleci/config.yml

This commit updates the node docker image used to build the project to version 18.16.0. It also simplifies the CircleCI workflow by removing unnecessary steps related to caching and restoring caches, and adding new parameters for nx/set-shas, allowing it to run with an on-hold workflow and specifying a main branch name. Finally, the production deployment is now filtered to run only on the main branch.

* build:
- Run lint with `yarn --silent run nx affected --base=$NX_BASE --head=$NX_HEAD --target=lint --parallel=3`
- Build server with `yarn --silent run nx build:server`

* build: dist/apps/my-app-server
feat: update yarn commands to npx in .circleci/config.yml

This commit updates the yarn commands in the deploy and build-app jobs in the .circleci/config.yml file to use npx instead of yarn. Specifically, it replaces "yarn --silent run" with "npx" for the nx commands in both jobs. This change simplifies the command syntax and ensures that the latest version of nx is used during the linting and building processes.

* build: add yarn cache and restore cache steps for CircleCI

This commit adds yarn caching to the CircleCI workflow to speed up dependencies installation, and also adds restore_cache steps to restore the cache when running jobs. Additionally, two new jobs are added to the workflow to lint and build the app using Nx affected. Finally, artifact storing is added to store the resulting build output.

* build: update yarn cache keys

This commit updates the keys used for caching dependencies. Specifically, the key prefix has been changed from "v1-dependency-cache" to "v3-dependency-cache". The new key is based on the checksum of the "yarn.lock" file rather than "package.json". This change ensures that the correct cache is used for each project and improves cache-hit rates. Additionally, this commit updates the cache key used for saving the cache.

* build server artifacts
chore: switch from npx to yarn for nx commands in CircleCI config.yml

* build: update yarn cache key to v4

This commit updates the yarn cache key from v3 to v4 in the CircleCI configuration file. The `save_cache` step and `key` value have been updated to reflect this change. Additionally, two `run` steps related to the `lint` and `build` targets have been commented out as they are not currently needed.

* build: update yarn dependency cache key and remove unused commands

The commit updates the yarn dependency cache key from v4 to v5, reflecting a change in the project's dependencies. Additionally, it removes two unused commands related to linting and building, which were commented out. The commit message follows the conventional commits specification.

* feat: update cache key from v5 to v6 in CircleCI configuration and remove nx lint fix from package.json

This commit updates the cache key from 'v5' to 'v6' in the CircleCI configuration file, specifically for the 'yarn_cache' key. Additionally, it removes the 'lint:fix' script from the 'package.json' file, which was previously used to run linting and formatting fixes across the entire workspace. Instead, the 'yarn nx affected' command is now used to target only affected projects for linting.

* fix: fix icon type declaration

* fix: type declaration for icons

* fix: remove log from server file

* fix: update declaration

* fix: temp

* update workspace root scripts (#168)

* DIA-173 bring down amount of nx-cloud cached runs (#169)

* feat(plp): prototype of jewelry plp

* feat: reworking logic to use re-normalized products

* refactor: updated the main diamonds api from post to get

* Feature/diamonds table (#38)

Co-authored-by: rconjoe <jcarlton@vrai.com>

* feat: common-ui slider (#170)

* feat: component naming pascal case (#174)

* feat: types for useDiamondData (#171)

* feat: types for useDiamondData

* feat: useDiamondData types

* feat: diamond data types

* feat: use-diamond-data fix thx sam

* feat: diamond-table get-server-side-props-ts-setup (#175)

* migration/06-07-2023 (#172)

* Feature/diamond table content dato (#176)

* feat: diamond-table get-server-side-props-ts-setup

* feat: diamond-table-content-dato

* feat: remove composite true

* feat: nx repair

* feat: translation

* feat: styling

* feat: active states filter nav

* feat: string ts bugfix

* feat: bug fixes

* feat: nit

* feat: diamondTable promo

* feat: styling

* feat: diamond table no result string

* feat: clean up

---------

Co-authored-by: rconjoe <jcarlton@vrai.com>

* feat: vd-1650 (#178)

* feat: vd-1650

* feat(vd-1650): added labels to fallback diamonds

* fix: rm log

* fix(diamond schema): added 'slug' property to schema (#180)

* refactor: updated the main diamonds api from post to get

* feat(plp): prototype of jewelry plp

* feat(plp api): outputting merged and reduced plp products

* feat(plp api): added support for er products in plp

* feat(plp filtering): added basic filtering to plp with basic ui

* feat(new libs): added product and diamond libs

* feat(lib): adding more files to product and diamond libs

* feat(product lib): adding more fn and files

* feat(product lib, diamond lib, plp api): code cleanup

* fix: merge conflict

* feat(pdp): refactored pdp logic to work with new normalized products

* style: Modify flexWrap property to flex in CatalogPage component

* build(lock): bump lock

* feat: rm test catalog pages

* feat(server product service): rm configuration endpoint used as a test

* feat: data cleanup

* feat(plp, pdp): adjusted code to support new normalized products

* feat: changed db collection back to 'products'

* feat(server): code cleanup

* feat: jsdocs

* refactor: cleanup and added some nest error logging

code cleanup and error logging

---------

Co-authored-by: div <samjdavidoff@gmail.com>
Co-authored-by: Bismark Frimpong <bismark@vrai.com>
Co-authored-by: rconjoe <jcarlton@vrai.com>
Co-authored-by: charles olivier <chla84@yahoo.fr>