feat: add plasma bash script and remove prelude script
This commit marks the addition of a new bash script file named 'plasma' and the deletion of an existing bash script named 'prelude'. The new 'plasma' script manages module sourcing and dependency checking for git repos and branches. On the other hand, 'prelude', an older version of environment manager script is removed.