feat: add dependencies checking module in plasma
This commit adds a new module named 'check_deps' in the .plasma directory. This new module helps in checking the existence of certain commands and installation of required dependencies on the system such as curl, gum, and offers support for various system platforms like MacOS, Linux (distinctly for Arch Linux and Debian), automatically installing them if not found. It also includes utility functions for upgrading gum, installing deno and installing or upgrading wmill based on the user's OS.