feat: add commit assist module
This newly added module includes the ability to generate a commit message based on the git log of staged files. It requires the `mods` package, and it will attempt to install this package if it's not already present. The script supports macOS, Arch Linux, and Debian systems, using Homebrew, paru, and apt package managers respectively.